requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch
torchvision
torchaudio
PySide6
requests
python-dotenv
tqdm
faiss-cpu
transformers
pyinstaller
tokenizers
Pillow
rarfile
pywin32
python-json-logger
zipp
markdown